Port program to PySide6
Specifically in line 76 of front, add loader.registerCustomWidget(plot_viewer).
On PySide6, the custom widget is not readily available to the QFormBuilder. So it
needs to be explicitly registered.

Add self.interface throughout the code because now, for some reason, the interface
needs to be called by its full reference.

Fix small typo on ui